replaced param tuning with new layout
